1. 支持URI中指定secretId和secretKey； 2. 支持三种类型的上传缓冲区。

+++ src/main/java/org/apache/hadoop/fs/Unit.java
@@ -4,9 +4,9 @@ public final class Unit {
     private Unit() {
     }
 
-    public static final int KB = 1024;
-    public static final int MB = 1024 * KB;
-    public static final int GB = 1024 * MB;
+    public static final long KB = 1024;
+    public static final long MB = 1024 * KB;
+    public static final long GB = 1024 * MB;
     public static final long TB = 1024 * GB;
     public static final long PB = 1024 * TB;
 }
